+++ oneflow/user/kernels/distributions/normal_kernel.h
@@ -51,8 +51,8 @@ class NormalKernel final : public user_op::OpKernel {
  private:
   void Compute(user_op::KernelComputeContext* ctx, user_op::OpKernelState* state) const override {
     user_op::Tensor* out = ctx->Tensor4ArgNameAndIndex("out", 0);
-    const float mean = ctx->Attr<float>("mean");
-    const double std = ctx->Attr<float>("std");
+    const double mean = ctx->Attr<double>("mean");
+    const double std = ctx->Attr<double>("std");
     int64_t elem_cnt = out->shape().elem_cnt();
     T* out_dptr = out->mut_dptr<T>();
     auto* normal_state = dynamic_cast<NormalKernelState*>(state);
